@@ -15,7 +15,7 @@ module Decidim::Admin
             description_ca: "La meva descripció",
             enable_omnipresent_banner: false,
             header_snippets: '<script>alert("Hello");</script>',
-            favicon: fixture_file_upload(File.new(Decidim::Dev.asset("icon.png")))
+            favicon: upload_test_file(Decidim::Dev.test_file("city.jpeg", "image/jpeg"))
           }
         }
       end
