@@ -3,11 +3,11 @@
 module Decidim
   class NewsletterMailer < ApplicationMailer
     helper Decidim::SanitizeHelper
+    helper Decidim::TranslationsHelper
+
     include Decidim::NewslettersHelper
 
     layout "decidim/newsletter_base"
-
-    add_template_helper Decidim::TranslationsHelper
 
     helper_method :cell
 
