@@ -28,7 +28,7 @@ module Decidim
     def render
       content_tag :li, class: link_wrapper_classes do
         output = [link_to(composed_label, url, link_options)]
-        output.push(@view.send(:simple_menu, @menu_item.submenu).render) if @menu_item.submenu
+        output.push(@view.send(:simple_menu, **@menu_item.submenu).render) if @menu_item.submenu
 
         safe_join(output)
       end
