@@ -16,7 +16,7 @@ module Decidim
         let(:phone) { "Y1fERVzL2F" }
         let(:document_number) { "123456780X" }
         let(:about) { "This is us." }
-        let(:avatar) { fixture_file_upload(File.open("spec/assets/avatar.jpg"), "image/jpeg") }
+        let(:avatar) { upload_test_file(Decidim::Dev.test_file("avatar.jpg", "image/jpeg")) }
 
         let(:form_params) do
           {
