@@ -466,10 +466,12 @@ shared_examples_for "an application with configurable env vars" do
     }
   end
 
+  # The logs settings have changed between Rails 6.0 abd 6.1 and this may be here
+  # https://github.com/rails/rails/commit/73079940111e8b85bf87953e5ef9fafeece5b5da
   let(:rails_off) do
     {
-      "Rails.logger.level" => 0,
-      "Rails.application.config.log_level" => "debug",
+      "Rails.logger.level" => 1,
+      "Rails.application.config.log_level" => "info",
       "Rails.application.config.action_controller.asset_host" => nil
     }
   end
