@@ -6,6 +6,8 @@ module Decidim
       # A command with all the business logic that updates an
       # existing initiative type.
       class UpdateInitiativeType < Decidim::Command
+        include ::Decidim::AttachmentAttributesMethods
+
         # Public: Initializes the command.
         #
         # initiative_type: Decidim::InitiativesType
@@ -39,7 +41,7 @@ module Decidim
         attr_reader :form, :initiative_type
 
         def attributes
-          result = {
+          {
             title: form.title,
             description: form.description,
             signature_type: form.signature_type,
@@ -55,10 +57,9 @@ module Decidim
             document_number_authorization_handler: form.document_number_authorization_handler,
             child_scope_threshold_enabled: form.child_scope_threshold_enabled,
             only_global_scope_enabled: form.only_global_scope_enabled
-          }
-
-          result[:banner_image] = form.banner_image unless form.banner_image.nil?
-          result
+          }.merge(
+            attachment_attributes(:banner_image)
+          )
         end
 
         def upate_initiatives_signature_type
