@@ -66,9 +66,8 @@ module Decidim
 
     validates :title, :description, :state, :signature_type, presence: true
     validates :hashtag,
-              uniqueness: true,
-              allow_blank: true,
-              case_sensitive: false
+              uniqueness: { allow_blank: true, case_sensitive: false }
+
     validate :signature_type_allowed
 
     scope :open, lambda {
