@@ -13,7 +13,7 @@ describe Decidim::Pages::Permissions do
     }
   end
   let(:page_component) { create :page_component }
-  let(:permission_action) { Decidim::PermissionAction.new(action) }
+  let(:permission_action) { Decidim::PermissionAction.new(**action) }
 
   context "when updating a page" do
     let(:action) do
