@@ -17,10 +17,7 @@ module Decidim::System
       }
     end
     let(:file) do
-      fixture_file_upload(
-        Decidim::Dev.test_file("city.jpeg", "image/jpeg"),
-        "image/jpeg"
-      )
+      upload_test_file(Decidim::Dev.test_file("city.jpeg", "image/jpeg"))
     end
     let(:application) { create(:oauth_application) }
     let(:organization) { application.organization }
